pyproject.toml
[tool.poetry]
name = "langchain-book"
version = "0.1.0"
description = ""
authors = ["arcstep <43801@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.0"
ipykernel = "^6.29.0"
jupyter = "^1.0.0"
python-dotenv = "^1.0.1"
docarray = "^0.32.0"
langchain-community = "^0.0.16"
langchain-core = "^0.1.16"
langchain-experimental = "^0.0.49"
langchain-openai = "^0.0.5"
httpx-sse = "^0.4.0"
fastapi = "^0.109.0"
langserve = "^0.0.41"
chromadb = "^0.4.22"
lancedb = "^0.5.1"
pandas = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
